+++ CHANGELOG.md
@@ -39,7 +39,7 @@
 ## [1.0.2] 2022-10-25
 ### Changes
 
-- Products pulled from Stripe
+- clients pulled from Stripe
   - Available for `Superusers`
 
 ## [1.0.1] 2022-10-18
@@ -50,7 +50,7 @@
 ## [1.0.0] 2022-10-18
 ### Stable Version
 
-- Products loaded from `templates` directory
+- clients loaded from `templates` directory
 - Stripe v3
   - `Stripe Credentials` loaded from `.env` file
  **[Soft UI Design](https://www.creative-tim.com/product/soft-ui-design-system?AFFILIATE=128200)** (Free Version) by `Creative-Tim`